Refactor stats tests
- add fixtures
- move helper methods into lib
